refactor : 패키지 구조 변경 및 컨벤션 수정
- domain 내 lotto 패키지 생성
- 사용하지 않는 import문 제거 및 공백, 개행 수정
