+++ src/main/java/lotto/Application.java
@@ -3,6 +3,7 @@ package lotto;
 import lotto.controller.LottoController;
 
 public class Application {
+
     public static void main(String[] args) {
         LottoController lottoController = new LottoController();
         lottoController.play();
