+++ src/main/java/lotto/controller/LottoController.java
@@ -5,10 +5,10 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import lotto.domain.Lotto;
-import lotto.domain.Lottos;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.Lottos;
 import lotto.domain.Money;
-import lotto.domain.Number;
+import lotto.domain.lotto.Number;
 import lotto.domain.Result;
 import lotto.view.InputView;
 import lotto.view.OutputView;
@@ -62,18 +62,13 @@ public class LottoController {
             isValid = validateNumbers(numbers);
         } while (!isValid);
 
-        return numbers.stream()
-            .map(String::valueOf)
-            .map(Number::new)
-            .collect(Collectors.toList());
+        return numbers.stream().map(String::valueOf).map(Number::new).collect(Collectors.toList());
     }
 
     private List<Integer> toList(String[] splitInput) {
         List<Integer> numbers = new ArrayList<>();
         try {
-            numbers = Arrays.stream(splitInput)
-                .map(Integer::parseInt)
-                .collect(Collectors.toList());
+            numbers = Arrays.stream(splitInput).map(Integer::parseInt).collect(Collectors.toList());
         } catch (NumberFormatException exception) {
         }
         return numbers;
