+++ src/main/java/lotto/domain/WinningPrice.java
@@ -5,10 +5,10 @@ import java.util.Arrays;
 public enum WinningPrice {
 
     Fail(0, false, 0),
-    Three(3, false,5_000),
-    Four(4, false,50_000),
-    Five(5,false, 1_500_000),
-    FiveAndBonus(5,true,30_000_000),
+    Three(3, false, 5_000),
+    Four(4, false, 50_000),
+    Five(5, false, 1_500_000),
+    FiveAndBonus(5, true, 30_000_000),
     All(6, false, 2_000_000_000);
 
     private final int count;
