+++ src/main/java/lotto/domain/lotto/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -6,6 +6,8 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
+
+import lotto.domain.WinningPrice;
 
 public class Lotto {
 
@@ -35,10 +37,7 @@ public class Lotto {
 
     Lotto(List<Integer> numbers) {
         validateNumbers(numbers);
-        this.numbers = numbers.stream()
-            .map(String::valueOf)
-            .map(Number::new)
-            .collect(Collectors.toSet());
+        this.numbers = numbers.stream().map(String::valueOf).map(Number::new).collect(Collectors.toSet());
     }
 
     public boolean contains(Number number) {
@@ -57,9 +56,7 @@ public class Lotto {
     }
 
     private static void validateDuplicate(List<Integer> numbers) {
-        int noDuplicateCount = (int)numbers.stream()
-            .distinct()
-            .count();
+        int noDuplicateCount = (int)numbers.stream().distinct().count();
 
         if (noDuplicateCount != SIZE) {
             throw new IllegalArgumentException(DUPLICATE_ERROR_MESSAGE);
