+++ src/main/java/lotto/domain/lotto/Number.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import java.util.Objects;
