+++ src/main/java/lotto/view/InputView.java
@@ -7,6 +7,7 @@ public class InputView {
     private static final String INPUT_MONEY_MESSAGE = "구입금액을 입력해 주세요.";
     private static final String INPUT_WINNING_NUMBERS_MESSAGE = "%n지난 주 당첨 번호를 입력해 주세요.%n";
     private static final String INPUT_BONUS_NUMBER_MESSAGE = "보너스 볼을 입력해 주세요.";
+
     private static final Scanner scanner = new Scanner(System.in);
 
     private InputView() {
