+++ src/main/java/lotto/view/OutputView.java
@@ -2,9 +2,9 @@ package lotto.view;
 
 import java.util.stream.Collectors;
 
-import lotto.domain.Lotto;
-import lotto.domain.Lottos;
-import lotto.domain.Number;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.Lottos;
+import lotto.domain.lotto.Number;
 import lotto.domain.Result;
 import lotto.domain.WinningPrice;
 
@@ -19,7 +19,8 @@ public class OutputView {
 
     public static void printLottos(Lottos lottos) {
         for (Lotto lotto : lottos.getLottos()) {
-            String join = lotto.getNumbers().stream()
+            String join = lotto.getNumbers()
+                .stream()
                 .map(Number::getValue)
                 .map(String::valueOf)
                 .collect(Collectors.joining(", "));
@@ -38,8 +39,8 @@ public class OutputView {
             if (value.isContainBonus()) {
                 bonusBallMessage = ", 보너스 볼 일치";
             }
-            System.out.printf("%d개 일치%s(%d원) - %d개%n",
-                value.getCount(),  bonusBallMessage, value.getPrice(), result.getCount(value));
+            System.out.printf("%d개 일치%s(%d원) - %d개%n", value.getCount(), bonusBallMessage, value.getPrice(),
+                result.getCount(value));
         }
     }
