+++ src/test/java/lotto/domain/MoneyTest.java
@@ -8,8 +8,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import lotto.domain.Money;
-
 public class MoneyTest {
 
     @Nested
