+++ src/test/java/lotto/domain/WinningPriceTest.java
@@ -6,8 +6,6 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-import lotto.domain.WinningPrice;
-
 public class WinningPriceTest {
 
     @Nested
